+++ vm/src/hint_processor/builtin_hint_processor/bootloader/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod types;
+pub(crate) mod vars;
