+++ vm/src/hint_processor/builtin_hint_processor/bootloader/types.rs
@@ -1,0 +1,41 @@
+use felt::Felt252;
+use serde::Deserialize;
+
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub struct BootloaderConfig {
+    pub simple_bootloader_program_hash: Felt252,
+    pub supported_cairo_verifier_program_hashes: Vec<Felt252>,
+}
+
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub enum PackedOutput {
+    Plain(Vec<Felt252>),
+    Composite(Vec<Felt252>),
+}
+
+impl PackedOutput {
+    // TODO: implement and define return type
+    pub fn elements_for_hash(&self) -> Vec<()> {
+        Default::default()
+    }
+}
+
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub struct FactTopology {}
+
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub struct Task {}
+
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub struct SimpleBootloaderInput {
+    pub fact_topologies_path: Option<String>,
+    pub single_page: bool,
+    pub tasks: Vec<Task>,
+}
+
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub struct BootloaderInput {
+    pub simple_bootloader_input: SimpleBootloaderInput,
+    pub bootloader_config: BootloaderConfig,
+    pub packed_outputs: Vec<PackedOutput>,
+}
