+++ vm/src/hint_processor/builtin_hint_processor/mod.rs
@@ -1,6 +1,7 @@
 pub mod bigint;
 pub mod blake2s_hash;
 pub mod blake2s_utils;
+mod bootloader;
 pub mod bootloader_hints;
 pub mod builtin_hint_processor_definition;
 pub mod cairo_keccak;
